pyproject.toml
[tool.poetry]
name = "ska-low-mccs-daq"
version = "0.2.0"
description = "The Monitoring, Control and Calibration Subsystem (MCCS) of the SKA Low telescope"
license = "BSD-3-Clause"
authors = ["MCCS team"]
readme = "README.md"
documentation = "https://developer.skatelescope.org/projects/ska-low-mccs-daq"

[[tool.poetry.source]]
name = 'ska-nexus'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'

[tool.poetry.dependencies]
# When updating this dependency, also update AAVS_SYSTEM_REF in the Dockerfile
aavs_system = { git = "https://gitlab.com/ska-telescope/aavs-system.git", rev = "498662646fcbb50c4995a1246f207852e2430006", subdirectory = "python" }
python = "~3.10"
future = "^0.18.2"
numpy = "^1.23.0"
pytango = "9.3.6"
PyYAML = "^6.0"
ska-tango-base = "^0.18.0"
ska-control-model = "0.3.1"
ska-low-mccs-common = "^0.6.1"
singleton-decorator = "1.0.0"
slacker = "0.14.0"
h5py = "3.7.0"
lockfile = "0.12.2"
matplotlib = "3.5.3"
grpcio = "1.51.1"
grpcio-tools = "1.51.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pylint = "^2.15.3"
pylint-junit = "^0.3.2"
pytest = "6.2.5"
pytest-bdd = "^6.0.1"
pytest-cov = "^3.0.0"
pytest-json-report = "^1.5.0"
pytest-mock = "^3.8.2"
pytest-repeat = "^0.9.1"
pytest-timeout = "^2.1.0"
pytest-xdist = "^2.5.0"
coverage = "^6.4.2"
black = "^22.6.0"
darglint = "^1.8.1"
flake8 = "^5.0.1"
flake8-formatter-junit-xml = "^0.0.6"
flake8-builtins = "^1.5.3"
flake8-docstrings = "^1.6.0"
flake8-use-fstring = "^1.4"
flake8-rst-docstrings = "^0.2.7"
mypy = "^0.971"
pep8-naming = "^0.13.1"
types-pkg-resources = "^0.1.3"
types-PyYAML = "^6.0.11"
isort = "^5.10.1"
ska-tango-testing = "^0.6.0"

[tool.poetry.scripts]
MccsDaqReceiver = "ska_low_mccs_daq.daq_receiver.daq_receiver_device:main"
MccsDaqServer = "ska_low_mccs_daq.gRPC_server.daq_grpc_server:main"

[tool.pytest.ini_options]
testpaths = "tests/"
addopts = "--json-report --json-report-file=build/reports/report.json --cov-report html:build/htmlcov --cov-report xml:build/reports/code-coverage.xml --cov-report term:skip-covered --cov=ska_low_mccs_daq --junitxml=build/reports/unit-tests.xml --verbose"
console_output_style = "progress"
junit_family = "legacy"

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
data_file = ".coverage/.coverage"
branch = true
source = ["ska_low_mccs_daq"]
omit = [
    "*/gRPC_server/generated_code/*",
    ]

[tool.coverage.report]
show_missing = false

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"
